requirements.txt
langchain
langchain_classic
langchain-core
langchain-community
langchain-groq
python-dotenv
streamlit
faiss-cpu
unstructured
beautifulsoup4
tiktoken
pypdf
sentence-transformers